# backend/swagger.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /products/feed:
    get:
      summary: Получить ленту товаров
      description: Возвращает отфильтрованную и отсортированную ленту товаров с предложениями
      operationId: get_products_feed_products_feed_get
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Смещение для пагинации
            default: 0
            title: Offset
          description: Смещение для пагинации
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Лимит товаров на странице
            default: 20
            title: Limit
          description: Лимит товаров на странице
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: null
            description: Поисковый запрос
            title: Search
          description: Поисковый запрос
        - name: category
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: null
            description: Фильтр по категории
            title: Category
          description: Фильтр по категории
        - name: merchant_ids
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: null
            description: Список ID магазинов через запятую
            title: Merchant Ids
          description: Список ID магазинов через запятую
        - name: sort_by
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SortOptions'
            description: Поле для сортировки
            default: price
          description: Поле для сортировки
        - name: user_lat
          in: query
          required: false
          schema:
            anyOf:
              - type: number
              - type: null
            description: Широта пользователя
            title: User Lat
          description: Широта пользователя
        - name: user_long
          in: query
          required: false
          schema:
            anyOf:
              - type: number
              - type: null
            description: Долгота пользователя
            title: User Long
          description: Долгота пользователя
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductFeedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /products/{ean}/offers:
    get:
      summary: Получить предложения для товара
      description: Возвращает все предложения для конкретного товара с ценами и расстояниями
      operationId: get_product_offers_products__ean__offers_get
      parameters:
        - name: ean
          in: path
          required: true
          schema:
            type: integer
            title: Ean
        - name: user_lat
          in: query
          required: true
          schema:
            type: number
            description: Широта пользователя
            title: User Lat
          description: Широта пользователя
          example: 55.7558
        - name: user_long
          in: query
          required: true
          schema:
            type: number
            description: Долгота пользователя
            title: User Long
          description: Долгота пользователя
          example: 37.6173
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductOffersListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /products/search:
    get:
      summary: Поиск товаров
      description: Поиск товаров по названию или категории
      operationId: search_products_products_search_get
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            description: Поисковый запрос
            title: Q
          description: Поисковый запрос
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Смещение для пагинации
            default: 0
            title: Offset
          description: Смещение для пагинации
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            description: Лимит товаров на странице
            default: 20
            title: Limit
          description: Лимит товаров на странице
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductFeedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProductFeedData-Output:
      properties:
        products:
          items:
            $ref: '#/components/schemas/ProductFeedItem'
          type: array
          title: Products
        total_count:
          type: integer
          title: Total Count
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
        - products
        - total_count
        - offset
        - limit
      title: ProductFeedData
    ProductFeedItem:
      properties:
        ean:
          type: integer
          title: Ean
        name:
          type: string
          title: Name
        category:
          type: string
          title: Category
        weight:
          type: number
          title: Weight
        offers:
          items:
            $ref: '#/components/schemas/ProductOffer'
          type: array
          title: Offers
        best_offer:
          anyOf:
            - $ref: '#/components/schemas/ProductOffer'
            - type: null
        min_price:
          type: number
          title: Min Price
        max_price:
          type: number
          title: Max Price
      type: object
      required:
        - ean
        - name
        - category
        - weight
        - offers
        - min_price
        - max_price
      title: ProductFeedItem
    ProductFeedResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/ProductFeedData-Output'
      type: object
      required:
        - message
        - data
      title: ProductFeedResponse
    ProductInfo:
      properties:
        ean:
          type: integer
          title: Ean
        name:
          type: string
          title: Name
        category:
          type: string
          title: Category
      type: object
      required:
        - ean
        - name
        - category
      title: ProductInfo
    ProductOffer:
      properties:
        sku_id:
          type: integer
          title: Sku Id
        price:
          type: number
          title: Price
        amount:
          type: integer
          title: Amount
        merchant_id:
          type: integer
          title: Merchant Id
        merchant_name:
          type: string
          title: Merchant Name
        stock_id:
          type: integer
          title: Stock Id
        stock_address:
          type: string
          title: Stock Address
        distance_km:
          anyOf:
            - type: number
            - type: null
          title: Distance Km
        stock_lat:
          type: number
          title: Stock Lat
        stock_long:
          type: number
          title: Stock Long
      type: object
      required:
        - sku_id
        - price
        - amount
        - merchant_id
        - merchant_name
        - stock_id
        - stock_address
        - stock_lat
        - stock_long
      title: ProductOffer
    ProductOffersListResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/ProductOffersResponse'
      type: object
      required:
        - message
        - data
      title: ProductOffersListResponse
    ProductOffersResponse:
      properties:
        product:
          $ref: '#/components/schemas/ProductInfo'
        offers:
          items:
            $ref: '#/components/schemas/ProductOffer'
          type: array
          title: Offers
        best_offer:
          anyOf:
            - $ref: '#/components/schemas/ProductOffer'
            - type: null
      type: object
      required:
        - product
        - offers
      title: ProductOffersResponse
...
    SortOptions:
      type: string
      enum:
        - price
        - distance
        - best_value
      title: SortOptions
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token